--- a/QML/BluePrintBoxes/BoxesFullScreenRepeater.qml
+++ b/QML/BluePrintBoxes/BoxesFullScreenRepeater.qml
@@ -5,14 +5,51 @@ import QtQuick.Window 2.15
 import "SingleBoxInstance"
 
 import org.bluePrintType.BPStatic 1.0
+import org.bluePrintType.OperationBox 1.0
 
 Repeater{
     id:boxesMainRepeaterID
     anchors.fill: parent
     model: pBluePrintPage.listOfBlueBoxes
-
     delegate: Loader{
-        source: appMainWindowID.defineGUIDeleage (blueBox_GUIType)
+        property COperationBox pBoxInstance: modelData
+
+        function checkVisibility(flickCntntX, flickCntntY, scrollViewWidth, scrollViewHeight,  itemScale , boxWidth, boxHeight, BoxPosX, boxPosY ) {
+            // Transform item position with scale
+            const scaledX = itemScale * BoxPosX
+            const scaledY = itemScale * boxPosY
+            const scaledW = itemScale * boxWidth
+            const scaledH = itemScale * boxHeight
+
+            const visibleLeft = flickCntntX
+            const visibleRight = flickCntntX + scrollViewWidth
+            const visibleTop = flickCntntY
+            const visibleBottom = flickCntntY + scrollViewHeight
+
+            const itemRight = scaledX + scaledW
+            const itemBottom = scaledY + scaledH
+
+            const isVisible = (
+                                itemRight > visibleLeft &&
+                                scaledX < visibleRight &&
+                                itemBottom > visibleTop &&
+                                scaledY < visibleBottom
+                                );
+            return isVisible;
+        }
+
+        property bool isItemNearVisibleArea : checkVisibility(scrollViewInappMainWindowID.flickableItem.contentX,
+                                                              scrollViewInappMainWindowID.flickableItem.contentY,
+                                                              scrollViewInappMainWindowID.width,
+                                                              scrollViewInappMainWindowID.height,
+                                                              bpPageItemID.scaleFactorOfApp,
+                                                              pBoxInstance.blueBox_width,
+                                                              pBoxInstance.blueBox_height,
+                                                              pBoxInstance.blueBox_xPos,
+                                                              pBoxInstance.blueBox_yPos
+                                                              );
+
+        source: isItemNearVisibleArea ?  appMainWindowID.defineGUIDeleage (blueBox_GUIType) : ""
         asynchronous: true
     }