🚀 **Massive Performance Boost in `QBluePrint`**

This small commit introduces a key optimization:
> 🧹 Blueprint boxes that are outside the visible range are now **destroyed** automatically.
Removing out-of-view flow lines could also be considered, but I'm letting that one be as is for now.
### 🔍 Before vs After (Tested on Core i7 + GTX 1070):

| Metric                         | Before                  | After                    |
|--------------------------------|-------------------------|--------------------------|
| Load Time (1000 nodes)         | ~40 seconds             | **< 1 second**           |
| Memory Usage (1000 nodes)      | > 600 MB                | **< 150 MB**             |
| Page Destruction / App Close   | > 20 seconds            | **Almost instant**       |

This leads to:
- **Over 80x faster** node loading
- **Significantly reduced** memory footprint
- **Instant cleanup** when closing the page or the application 

If user is working with large number of nodes in `QBluePrint`, this should make a *huge* difference.
this commit Closes #25